+++ surfsense_backend/app/services/connector_service.py
@@ -16,8 +16,8 @@ from app.db import (
     SearchSourceConnector,
     SearchSourceConnectorType,
 )
-from app.retriver.chunks_hybrid_search import ChucksHybridSearchRetriever
-from app.retriver.documents_hybrid_search import DocumentHybridSearchRetriever
+from app.retriever.chunks_hybrid_search import ChucksHybridSearchRetriever
+from app.retriever.documents_hybrid_search import DocumentHybridSearchRetriever
 
 
 class ConnectorService:
